+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+#### Version 1.0.7 (2022-03-08)
+* Update methods to be conform with libjs 1.4.0
+
+#### Version 1.0.6 (2022-03-01)
+* Add default value to curve type in the address 
+
 #### Version 1.0.5 (2022-01-15)
 * Rename node field in CrossValidationStamp with 'nodePublicKey'
 * Add flutter_lints